requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0

pydantic>=2.0
pydantic-settings
python-dotenv

python-jose[cryptography]
passlib[bcrypt]

pytest
httpx